- Se agregaron los Binding de (Descripcion, Existencia, Costo y ValorInventario), y se agregó una nueva migracion y base de datos (Totalmente nuevas).

+++ Entidades/Articulos.cs
@@ -12,8 +12,8 @@ namespace PrimerParcial_JoseLuis.Entidades
         [Key]
         public int IdArticulo { get; set; }
         public string Descripcion { get; set; }
-        public int Existencia { get; set; }
-        public int Costo { get; set; }
-        public int ValorInventario { get; set; }
+        public double Existencia { get; set; }
+        public double Costo { get; set; }
+        public double ValorInventario { get; set; }
     }
 }
